Moving state to Game component
**Notes:**
So we previously moved the state up from the squares to the board. This worked great, until we wanted to track the history of all moves. Since that's best managed by the Game component, we've moved our data up there.

Like before, we switch this.state with this.props. As a reminder, properties are arbitrary variables used by a component. We move over our handleClick to the Game component, which is passed down to the Board component. The Board component then passes it to each square. When the user clicks, it's therefore propagated to Square => Board => Game.