Add key to array components
**NOTES:**
React doesn't re-render components until it has to.  The problem is: what if you have a list of components in an array? How does it know when to re-render?

That's where `key` comes into play. `key` is a reserved word used to help React manage exactly that. When a list is modified, React will check the `key` of all components in the array. If there's a key in the list that didn't exist before, React creates a new component. If there's *not* a key in the list that existed before, React destroys the component. If the key exists in both but it's moved to a different position, React will move the component. State is maintained.

By default, React will give each component its position in the list as its key. This is Not Great as it creates issues when moving, creating, or deleting items. As a result, React throws a warning in the dev console when this happens.

For our purposes, though, that's okay! We're going to assign the move number (the same thing as the array index) as the key. Since the entire point of a history is that it *is not modified at any point*, that works out fine. Generally, though: don't do this.