Finish game history
**NOTES:**
Not a lot to add here notes-wise, just a reiteration of previous ideas. The only notable thing is that to allow the player to go back in time and try different moves, we're using history = this.state.history.slice(0, this.state.stepNumber + 1). If you go back in time and then back a new move, the history after that point will be wiped.